* Restructuring the solution, moving code across the projects
