+++ R4MVC.sln
@@ -1,8 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26430.6
+VisualStudioVersion = 15.0.26430.12
 MinimumVisualStudioVersion = 10.0.40219.1
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".solution", ".solution", "{1E8F7AFD-5175-406A-9736-817C75DA1142}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{30626644-0E8A-4610-A0A9-274E91F1A037}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{E11DD8FA-FBC7-490C-88B7-15DB073FD903}"
@@ -12,6 +17,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R4MvcHostApp", "src\R4MvcHostApp\R4MvcHostApp.csproj", "{60691734-BD6C-4BED-8FAE-77E324C76574}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "R4Mvc.Test", "test\R4Mvc.Test\R4Mvc.Test.csproj", "{CFE51159-0850-415A-BAD0-FBFF607E772F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "R4Mvc.Tools", "src\R4Mvc.Tools\R4Mvc.Tools.csproj", "{1673B75E-E97D-4770-A1E3-86CF3C50A5E8}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -31,6 +38,10 @@ Global
 		{CFE51159-0850-415A-BAD0-FBFF607E772F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CFE51159-0850-415A-BAD0-FBFF607E772F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CFE51159-0850-415A-BAD0-FBFF607E772F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1673B75E-E97D-4770-A1E3-86CF3C50A5E8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1673B75E-E97D-4770-A1E3-86CF3C50A5E8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1673B75E-E97D-4770-A1E3-86CF3C50A5E8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1673B75E-E97D-4770-A1E3-86CF3C50A5E8}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -39,5 +50,6 @@ Global
 		{0DF253A4-4093-4299-98A1-D28175DFEE34} = {E11DD8FA-FBC7-490C-88B7-15DB073FD903}
 		{60691734-BD6C-4BED-8FAE-77E324C76574} = {E11DD8FA-FBC7-490C-88B7-15DB073FD903}
 		{CFE51159-0850-415A-BAD0-FBFF607E772F} = {30626644-0E8A-4610-A0A9-274E91F1A037}
+		{1673B75E-E97D-4770-A1E3-86CF3C50A5E8} = {E11DD8FA-FBC7-490C-88B7-15DB073FD903}
 	EndGlobalSection
 EndGlobal
