+++ src/R4Mvc.Tools/Program.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace R4Mvc.Tools
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+}
