+++ src/R4Mvc.Tools/R4MvcGenerator.cs
@@ -5,10 +5,10 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-using R4Mvc.Extensions;
-using R4Mvc.Services;
+using R4Mvc.Tools.Extensions;
+using R4Mvc.Tools.Services;
 
-namespace R4Mvc
+namespace R4Mvc.Tools
 {
 	public class R4MvcGenerator
 	{
