+++ src/R4Mvc.Tools/Services/ControllerRewriterService.cs
@@ -6,9 +6,9 @@ using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
 
-using R4Mvc.Extensions;
+using R4Mvc.Tools.Extensions;
 
-namespace R4Mvc.Services
+namespace R4Mvc.Tools.Services
 {
 	public class ControllerRewriterService : IControllerRewriterService
 	{
