+++ src/R4Mvc/R4MVCExcludeAttribute.cs
@@ -2,11 +2,11 @@
 
 namespace Microsoft.AspNetCore.Mvc
 {
-	/// <summary>
-	/// Force exclusion of a class for R4MVC code generation
-	/// </summary>
-	[AttributeUsage(AttributeTargets.Class)]
-	public class R4MVCExcludeAttribute : Attribute
-	{
-	}
+    /// <summary>
+    /// Force exclusion of a class for R4MVC code generation
+    /// </summary>
+    [AttributeUsage(AttributeTargets.Class)]
+    public class R4MvcExcludeAttribute : Attribute
+    {
+    }
 }
