+++ test/R4Mvc.Test/Services/SettingsTests.cs
@@ -1,5 +1,5 @@
 ﻿using System.IO;
-using R4Mvc.Services;
+using R4Mvc.Tools.Services;
 using Xunit;
 
 namespace R4Mvc.Test.Services
